fix-post-prod (#21)
* fix(TransactionForm): select input content on focus

* feat(capacitor): add status bar plugin and configuration

The changes in this commit add the Capacitor Status Bar plugin and configure it for the Android platform. The key changes are:

1. Add the `@capacitor/status-bar` dependency to the `package.json` and `yarn.lock` files.
2. Include the `capacitor-status-bar` project in the `capacitor.settings.gradle` file.
3. Configure the Status Bar plugin in the `capacitor.config.json` file, including setting the overlay, style, and background color.
4. Add the `capacitor-status-bar` dependency to the `capacitor.build.gradle` file.

These changes will enable the app to customize the status bar appearance on Android devices, improving the overall user experience.

* feat(transactions): add category icon property

The changes in this commit add a new `icon` property to the `Transaction` interface and the `Category` model in the database. This allows storing and displaying an icon associated with each category, providing a visual representation of the category type.

The changes include:

- Adding the `icon` property to the `Transaction` interface in `transactions-store.ts`.
- Adding the `icon` column to the `categories` table in the database schema.
- Updating the `Category` type in `categories-store.ts` to include the `icon` property.
- Updating the `CategoriesPage.vue` component to handle the new `icon` property when creating, editing, and displaying categories.
- Adding French translations for the new `icon` label and a set of predefined icons in `fr-FR/index.ts`.

These changes enhance the user experience by providing a visual cue for the category type, making it easier for users to quickly identify and understand their transactions.

* feat: Bump version to 0.1.0 and update build config

This commit bumps the version to 0.1.0 in both the Android app and the
package.json file. It also updates the Android app's build configuration
to use the new version code and name.